+++ src/api/index.ts
@@ -1,10 +1,12 @@
 import * as fs from "./fs";
-import * as layout from "./layout";
 import * as replDb from "./replDb";
 import * as me from "./me";
-import * as theme from "./theme";
+import * as themes from "./theme";
+import * as messages from "./messages";
+import * as data from "./data";
+import * as session from "./session";
 
-export { layout, fs, replDb, me, theme };
+export { fs, replDb, me, themes, messages, data, session };
 
 // deprecate this after migrating existing extensions
 export * from "./fs";