+++ src/types/index.ts
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 import { ChangeSpec } from "@codemirror/state";
 
 
@@ -5,18 +6,17 @@ export type Pane = {
   type: string;
   id: string;
 };
+=======
+/*****************************************************************
+ * * FS (Filesystem) Types
+ *****************************************************************/
+>>>>>>> origin/main
 
 /**
- * Enumeration of file types.
+ * A Filesystem node type
  */
-export enum FileType {
-  /**
-   * A regular file.
-   */
+export enum FsNodeType {
   File = "FILE",
-  /**
-   * A directory/folder
-   */
   Directory = "DIRECTORY",
 }
 
@@ -26,17 +26,11 @@ export enum FileType {
  * does not expose the node's content/children
  */
 export interface FsNode {
-  /**
-   * Full path of the node relative to the root
-   */
   path: string;
-
-  /**
-   * node is a file
-   */
-  type: FileType;
+  type: FsNodeType;
 }
 
+<<<<<<< HEAD
 export interface MoveEvent {
   eventType: 'MOVE';
   node: FsNode;
@@ -121,112 +115,134 @@ export interface FloatingPaneGroup {
   panes: Array<PaneId>;
   activeIndex: number;
   rect: Rect;
+=======
+export interface WatchFileWatchers {
+  onChange: WatchFileWatcherOnChange;
+  onError: WatchFileWatcherOnError;
+  onMoveOrDelete: WatchFileWatcherOnMoveOrDelete;
 }
 
+export interface WriteChangeArgs {
+  from: number;
+  to: number;
+  insert: string;
+}
+
+export interface WatchTextFileWatchers {
+  onReady: WatchTextFileWatcherOnReady;
+  onChange: WatchTextFileWatcherOnChange;
+  onError: WatchTextFileWatcherOnError;
+  onMoveOrDelete: WatchTextFileWatcherOnMoveOrDelete;
+}
+
+export interface DirectoryChildNode {
+  filename: string;
+  type: FsNodeType;
+>>>>>>> origin/main
+}
+
+/*****************************************************************
+ * * Theme Types
+ *****************************************************************/
 export type CssColor = string;
 
-export interface ThemeValues {
-  global: {
-    backgroundRoot: CssColor;
-    backgroundDefault: CssColor;
-    backgroundHigher: CssColor;
-    backgroundHighest: CssColor;
-    backgroundOverlay: CssColor;
-    foregroundDefault: CssColor;
-    foregroundDimmer: CssColor;
-    foregroundDimmest: CssColor;
-    outlineDimmest: CssColor;
-    outlineDimmer: CssColor;
-    outlineDefault: CssColor;
-    outlineStronger: CssColor;
-    outlineStrongest: CssColor;
-    accentPrimaryDimmest: CssColor;
-    accentPrimaryDimmer: CssColor;
-    accentPrimaryDefault: CssColor;
-    accentPrimaryStronger: CssColor;
-    accentPrimaryStrongest: CssColor;
-    accentPositiveDimmest: CssColor;
-    accentPositiveDimmer: CssColor;
-    accentPositiveDefault: CssColor;
-    accentPositiveStronger: CssColor;
-    accentPositiveStrongest: CssColor;
-    accentNegativeDimmest: CssColor;
-    accentNegativeDimmer: CssColor;
-    accentNegativeDefault: CssColor;
-    accentNegativeStronger: CssColor;
-    accentNegativeStrongest: CssColor;
-    redDimmest: CssColor;
-    redDimmer: CssColor;
-    redDefault: CssColor;
-    redStronger: CssColor;
-    redStrongest: CssColor;
-    orangeDimmest: CssColor;
-    orangeDimmer: CssColor;
-    orangeDefault: CssColor;
-    orangeStronger: CssColor;
-    orangeStrongest: CssColor;
-    yellowDimmest: CssColor;
-    yellowDimmer: CssColor;
-    yellowDefault: CssColor;
-    yellowStronger: CssColor;
-    yellowStrongest: CssColor;
-    limeDimmest: CssColor;
-    limeDimmer: CssColor;
-    limeDefault: CssColor;
-    limeStronger: CssColor;
-    limeStrongest: CssColor;
-    greenDimmest: CssColor;
-    greenDimmer: CssColor;
-    greenDefault: CssColor;
-    greenStronger: CssColor;
-    greenStrongest: CssColor;
-    tealDimmest: CssColor;
-    tealDimmer: CssColor;
-    tealDefault: CssColor;
-    tealStronger: CssColor;
-    tealStrongest: CssColor;
-    blueDimmest: CssColor;
-    blueDimmer: CssColor;
-    blueDefault: CssColor;
-    blueStronger: CssColor;
-    blueStrongest: CssColor;
-    blurpleDimmest: CssColor;
-    blurpleDimmer: CssColor;
-    blurpleDefault: CssColor;
-    blurpleStronger: CssColor;
-    blurpleStrongest: CssColor;
-    purpleDimmest: CssColor;
-    purpleDimmer: CssColor;
-    purpleDefault: CssColor;
-    purpleStronger: CssColor;
-    purpleStrongest: CssColor;
-    magentaDimmest: CssColor;
-    magentaDimmer: CssColor;
-    magentaDefault: CssColor;
-    magentaStronger: CssColor;
-    magentaStrongest: CssColor;
-    pinkDimmest: CssColor;
-    pinkDimmer: CssColor;
-    pinkDefault: CssColor;
-    pinkStronger: CssColor;
-    pinkStrongest: CssColor;
-    greyDimmest: CssColor;
-    greyDimmer: CssColor;
-    greyDefault: CssColor;
-    greyStronger: CssColor;
-    greyStrongest: CssColor;
-    brownDimmest: CssColor;
-    brownDimmer: CssColor;
-    brownDefault: CssColor;
-    brownStronger: CssColor;
-    brownStrongest: CssColor;
-    black: CssColor;
-    white: CssColor;
-  };
-  editor: {
-    //   syntaxHighlighting: Array<SyntaxHighlightingSpec>;
-    syntaxHighlighting: any;
-  };
+export interface ThemeValuesGlobal {
+  __typename?: string;
+  backgroundRoot: CssColor;
+  backgroundDefault: CssColor;
+  backgroundHigher: CssColor;
+  backgroundHighest: CssColor;
+  backgroundOverlay: CssColor;
+  foregroundDefault: CssColor;
+  foregroundDimmer: CssColor;
+  foregroundDimmest: CssColor;
+  outlineDimmest: CssColor;
+  outlineDimmer: CssColor;
+  outlineDefault: CssColor;
+  outlineStronger: CssColor;
+  outlineStrongest: CssColor;
+  accentPrimaryDimmest: CssColor;
+  accentPrimaryDimmer: CssColor;
+  accentPrimaryDefault: CssColor;
+  accentPrimaryStronger: CssColor;
+  accentPrimaryStrongest: CssColor;
+  accentPositiveDimmest: CssColor;
+  accentPositiveDimmer: CssColor;
+  accentPositiveDefault: CssColor;
+  accentPositiveStronger: CssColor;
+  accentPositiveStrongest: CssColor;
+  accentNegativeDimmest: CssColor;
+  accentNegativeDimmer: CssColor;
+  accentNegativeDefault: CssColor;
+  accentNegativeStronger: CssColor;
+  accentNegativeStrongest: CssColor;
+  redDimmest: CssColor;
+  redDimmer: CssColor;
+  redDefault: CssColor;
+  redStronger: CssColor;
+  redStrongest: CssColor;
+  orangeDimmest: CssColor;
+  orangeDimmer: CssColor;
+  orangeDefault: CssColor;
+  orangeStronger: CssColor;
+  orangeStrongest: CssColor;
+  yellowDimmest: CssColor;
+  yellowDimmer: CssColor;
+  yellowDefault: CssColor;
+  yellowStronger: CssColor;
+  yellowStrongest: CssColor;
+  limeDimmest: CssColor;
+  limeDimmer: CssColor;
+  limeDefault: CssColor;
+  limeStronger: CssColor;
+  limeStrongest: CssColor;
+  greenDimmest: CssColor;
+  greenDimmer: CssColor;
+  greenDefault: CssColor;
+  greenStronger: CssColor;
+  greenStrongest: CssColor;
+  tealDimmest: CssColor;
+  tealDimmer: CssColor;
+  tealDefault: CssColor;
+  tealStronger: CssColor;
+  tealStrongest: CssColor;
+  blueDimmest: CssColor;
+  blueDimmer: CssColor;
+  blueDefault: CssColor;
+  blueStronger: CssColor;
+  blueStrongest: CssColor;
+  blurpleDimmest: CssColor;
+  blurpleDimmer: CssColor;
+  blurpleDefault: CssColor;
+  blurpleStronger: CssColor;
+  blurpleStrongest: CssColor;
+  purpleDimmest: CssColor;
+  purpleDimmer: CssColor;
+  purpleDefault: CssColor;
+  purpleStronger: CssColor;
+  purpleStrongest: CssColor;
+  magentaDimmest: CssColor;
+  magentaDimmer: CssColor;
+  magentaDefault: CssColor;
+  magentaStronger: CssColor;
+  magentaStrongest: CssColor;
+  pinkDimmest: CssColor;
+  pinkDimmer: CssColor;
+  pinkDefault: CssColor;
+  pinkStronger: CssColor;
+  pinkStrongest: CssColor;
+  greyDimmest: CssColor;
+  greyDimmer: CssColor;
+  greyDefault: CssColor;
+  greyStronger: CssColor;
+  greyStrongest: CssColor;
+  brownDimmest: CssColor;
+  brownDimmer: CssColor;
+  brownDefault: CssColor;
+  brownStronger: CssColor;
+  brownStrongest: CssColor;
+  black: CssColor;
+  white: CssColor;
 }
 
 export enum ColorScheme {
@@ -234,15 +250,21 @@ export enum ColorScheme {
   Dark = "dark",
 }
 
-export interface Theme {
-  id: string;
-  description: string;
-  values: ThemeValues;
+export interface CustomTheme {
+  author: User;
   colorScheme: ColorScheme;
-  name: string;
-  isOfficial: boolean;
+  hasUnpublishedChanges: boolean;
+  id: number;
+  isCurrentUserThemeAuthor: boolean;
+  isInstalledByCurrentUser: boolean;
+  latestThemeVersion: ThemeVersion;
+  numInstalls?: number;
+  slug?: string;
+  status?: "public" | "private";
+  title?: string;
 }
 
+<<<<<<< HEAD
 export type ExtensionPortAPI = {
   // misc
   handshake: (args: { permissions: Array<string> }) => void;
@@ -263,52 +285,282 @@ export type ExtensionPortAPI = {
   copyFile: (path: string, to: string) => Promise<{ error: string | null }>;
   watchFile: (path: string, watcher: WatchFileListeners) => () => void;
   watchTextFile: (path: string, watcher: WatchTextFileListeners) => () => void;
+=======
+export interface ThemeSyntaxHighlightingTag {
+  __typename: string;
+  name: string;
+  modifiers: null | Array<string>;
+}
+>>>>>>> origin/main
 
-  // replDb
-  setReplDbValue: (key: string, value: string) => Promise<void>;
-  getReplDbValue: (key: string) => Promise<string | null>;
-  listReplDbKeys: (
-    prefix: string
-  ) => Promise<{ keys: string[] } | { error: string }>;
-  deleteReplDbKey: (key: string) => Promise<void>;
+export interface ThemeSyntaxHighlightingModifier {
+  textDecoration?: string;
+  fontSize?: string;
+  fontWeight?: string;
+  fontStyle?: string;
+  color?: string;
+}
 
-  // layout
-  isPaneTypeVisible: (paneType: string) => Promise<boolean>;
-  findPaneByType: (paneType: string) => Promise<{
-    paneId: string;
-    isHidden: boolean;
-    isDialog: boolean;
-    type: "tile" | "floating";
-    data: any;
-  } | null>;
-  selectTab: (paneId: string) => Promise<void>;
-  insertFloatingPaneIfNotExist(pane: Pane): Promise<void>;
-  removeFloatingPanesByType(paneType: string): Promise<void>;
-  getLayoutState(): Promise<LayoutData>;
-  setLayoutState(state: LayoutData): Promise<void>;
+export interface ThemeEditorSyntaxHighlighting {
+  __typename: string;
+  tags: Array<ThemeSyntaxHighlightingTag>;
+  values: ThemeSyntaxHighlightingModifier;
+}
 
-  activatePane: () => Promise<void>;
-  // theme
-  getCurrentTheme: () => Promise<Theme>;
-  onThemeChange: (callback: (theme: Theme) => void) => Promise<() => void>;
+export interface ThemeValuesEditor {
+  editor: Array<ThemeEditorSyntaxHighlighting>;
+}
 
-  // jets (will be deprecated)
+export interface ThemeValues {
+  __typename?: string;
+  editor: ThemeValuesEditor;
+  global: ThemeValuesGlobal;
+}
 
-  // graphql
-  queryGraphql(args: { query: string; variables?: Record<string, any> }): any;
-  mutateGraphql(args: {
-    mutation: string;
-    variables?: Record<string, any>;
-  }): any;
+export interface ThemeVersion {
+  __typename?: string;
+  id: number;
+  hue: number;
+  lightness: number;
+  saturation: number;
+  timeUpdated?: string;
+  description?: string;
+  customTheme?: CustomTheme;
+  values?: ThemeValues;
+}
 
-  // eval
-  eval(code: string): any;
+/*****************************************************************
+ * * Data Input/Output Types
+ *****************************************************************/
 
-  filePath: string;
-};
+export interface UserDataInclusion {
+  includeSocialData?: boolean;
+  includeRoles?: boolean;
+}
+
+export interface ReplDataInclusion {
+  includeSocialData?: boolean;
+  includeComments?: boolean;
+  includeOwner?: boolean;
+  includeMultiplayers?: boolean;
+}
+
+export type GraphResponse<T> = Promise<T | never>;
+
+export type ReplQueryOutput = GraphResponse<{ repl: Repl }>;
+export type UserByUsernameQueryOutput = GraphResponse<{ userByUsername: User }>;
+export type UserQueryOutput = GraphResponse<{ user: User }>;
+
+/*****************************************************************
+ * * GraphQL Types
+ *****************************************************************/
+
+// User
+export interface User {
+  id: number;
+  username: string;
+  image: string;
+  bio?: string;
+
+  // SocialUserData fragment
+  url?: string;
+  socials?: Array<UserSocial>;
+  firstName?: string;
+  lastName?: string;
+  displayName?: string;
+  fullName?: string;
+  followCount?: number;
+  followerCount?: number;
+
+  // RolesUserData fragment
+  roles?: Array<UserRole>;
+}
+
+export interface UserSocial {
+  id: number;
+  url: string;
+  type: UserSocialType;
+}
+export enum UserSocialType {
+  twitter = "twitter",
+  github = "github",
+  linkedin = "linkedin",
+  website = "website",
+  youtube = "youtube",
+  twitch = "twitch",
+  facebook = "facebook",
+  discord = "discord",
+}
+export interface UserRole {
+  id: number;
+  name: string;
+  key: string;
+  tagline: string;
+}
+
+// Repl
+export interface Repl {
+  id: string;
+  url: string;
+  title: string;
+  description: string;
+  timeCreated: string;
+  slug: string;
+  isPrivate: boolean;
+
+  // SocialReplData fragment
+  likeCount?: number;
+  publicForkCount?: number;
+  runCount?: number;
+  commentCount?: number;
+  tags?: Array<Tag>;
+
+  // CommentsReplData fragment
+  comments?: {
+    items?: Array<ReplComment>;
+  };
+
+  // OwnerData fragment
+  owner?: {
+    __typename: string;
+    id: number;
+    username: string;
+    image: string;
+    bio?: string;
+    description?: string;
+  };
+
+  // MultiplayersData fragment
+  multiplayers?: Array<User>;
+}
+
+export interface Tag {
+  id: string;
+  isOfficial: boolean;
+}
+export interface ReplComment {
+  id: number;
+  body: string;
+  user: User;
+}
+
+/*****************************************************************
+ * * Miscalleneous / React Types / Function Args
+ *****************************************************************/
 
 export enum HandshakeStatus {
   Ready = "ready",
   Error = "error",
   Loading = "loading",
 }
+
+export interface StrError {
+  error: string;
+}
+export interface NullableStrError {
+  error: string | null;
+}
+
+export interface TextFileWatcherReadyArgs {
+  initialContent: string;
+  version: number;
+  writeChange: (writeChangeArgs: WriteChangeArgs) => Promise<void>;
+}
+export interface TextFileWatcherOnChangeArgs {
+  latestContent: string;
+  version: number;
+  changeSource: string;
+  changes: any; // TODO fix
+}
+export interface OnMoveOrDeleteArgs {
+  eventType: "MOVE" | "DELETE";
+  node: FsNode;
+}
+export type OnActiveFileChangeCallback = (file: string) => void;
+export type WatchFileWatcherOnChange = (newContent: string) => void;
+export type WatchFileWatcherOnError = (error: string) => void;
+export type WatchFileWatcherOnMoveOrDelete = (args: OnMoveOrDeleteArgs) => void;
+export type WatchTextFileWatcherOnReady = (
+  readyArgs: TextFileWatcherReadyArgs
+) => void;
+export type WatchTextFileWatcherOnChange = (
+  changeArgs: TextFileWatcherOnChangeArgs
+) => void;
+export type WatchTextFileWatcherOnError = (error: string) => void;
+export type WatchTextFileWatcherOnMoveOrDelete = (
+  args: OnMoveOrDeleteArgs
+) => void;
+export type HandshakeOuput = Promise<null | (() => void)>;
+export type OnThemeChangeValuesCallback = (values: ThemeValuesGlobal) => void;
+export type OnThemeChangeCallback = (theme: ThemeVersion) => void;
+
+/*****************************************************************
+ * * Extension Port Wrapper
+ *****************************************************************/
+
+export type ExtensionPortAPI = {
+  handshake: () => void;
+
+  // fs Module
+  readFile: (
+    path: string,
+    encoding: "utf8" | "binary" | null
+  ) => Promise<{ content: string } | StrError>;
+  writeFile: (
+    path: string,
+    content: string | Blob
+  ) => Promise<{ success: boolean } | StrError>;
+  readDir: (path: string) => Promise<{
+    children: Array<DirectoryChildNode>;
+    error: string;
+  }>;
+  createDir: (path: string) => Promise<{} | StrError>;
+  deleteFile: (path: string) => Promise<{} | StrError>;
+  deleteDir: (path: string) => Promise<{} | StrError>;
+  move: (path: string, to: string) => Promise<NullableStrError>;
+  copyFile: (path: string, to: string) => Promise<NullableStrError>;
+  watchFile: (path: string, watcher: WatchFileWatchers) => () => void;
+  watchTextFile: (path: string, watcher: WatchTextFileWatchers) => () => void;
+
+  // replDb Module
+  setReplDbValue: (key: string, value: string) => Promise<void>;
+  getReplDbValue: (key: string) => NullableStrError;
+  listReplDbKeys: (prefix: string) => Promise<{ keys: string[] } | StrError>;
+  deleteReplDbKey: (key: string) => Promise<void>;
+
+  activatePane: () => Promise<void>;
+
+  // theme
+  getCurrentThemeValues: () => Promise<ThemeValuesGlobal>;
+  onThemeChangeValues: (
+    callback: OnThemeChangeValuesCallback
+  ) => Promise<() => void>;
+  getCurrentTheme: () => Promise<ThemeVersion>;
+  onThemeChange: (
+    callback: (theme: ThemeVersion) => void
+  ) => Promise<() => void>;
+
+  filePath: string;
+
+  // messages Module
+  showConfirm: (text: string, length?: number) => string;
+  showError: (text: string, length?: number) => string;
+  showNotice: (text: string, length?: number) => string;
+  showWarning: (text: string, length?: number) => string;
+  hideMessage: (id: string) => void;
+  hideAllMessages: () => void;
+
+  // data Module
+  currentUser: (args: UserDataInclusion) => UserQueryOutput;
+  userById: (args: { id: string } & UserDataInclusion) => UserQueryOutput;
+  userByUsername: (
+    args: { username: string } & UserDataInclusion
+  ) => UserByUsernameQueryOutput;
+  currentRepl: (args: ReplDataInclusion) => ReplQueryOutput;
+  replById: (args: { id: string } & ReplDataInclusion) => ReplQueryOutput;
+  replByUrl: (args: { url: string } & ReplDataInclusion) => ReplQueryOutput;
+
+  // session Module
+  watchActiveFile: (callback: (path: string) => void) => () => void;
+  getActiveFile: () => Promise<string | null>;
+};